durty work with opengl and window size

+++ src/Core/Renderer.h
@@ -2,7 +2,6 @@
 #define RENDERER_H
 #include "graphics/Opengl.h"
 
-
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -11,8 +10,13 @@ typedef struct Renderer {
 	OpenGL* opengl;
 } Renderer;
 
+// TODO: this function should be call when window is resizing (unsupported yet)
 void render_on_resize(OpenGL* gl, int width, int height);
-void render_update(float dt);
+
+// i dont know if its function needed i can render with interpolate time, maybe dt need for physics calls
+void render_update(double dt);
+
+// makes window current, pain and swap buffers, maybe its possiblt to divide into 2 functions
 void render(OpenGL* gl);
 
 #ifdef __cplusplus
